Optimize confirmation popups
