+++ script/ui/MainMenu.cs
@@ -3,31 +3,14 @@ using System;
 
 public partial class MainMenu : Node
 {
-	private ConfirmationDialog _confirmationPopup;
-
-	public override void _Process(double delta)
-	{
-		if(_confirmationPopup != null)
-		{
-			CheckForUserInputInPopup();
-		}
-	}
+	private ConfirmationPopup _confirmationPopup;
 
 	private void OnExitButtonPress()
 	{
-		_confirmationPopup = (ConfirmationDialog)GD.Load<PackedScene>("res://scene//ui//ConfirmationDialog.tscn").Instantiate();
+		_confirmationPopup = (ConfirmationPopup)GD.Load<PackedScene>("res://scene//ui//ConfirmationPopup.tscn").Instantiate();
 		_confirmationPopup.Init("Willst du wirklich das Spiel verlassen?", "Spiel verlassen");
+		_confirmationPopup.Confirmed += () => GetTree().Quit();
 		AddChild(_confirmationPopup);
-
-	}
-
-	private void CheckForUserInputInPopup()
-	{
-		if (_confirmationPopup.Confirmed)
-		{
-			_confirmationPopup.QueueFree();
-			GetTree().Quit();
-		}
 	}
 
 	public void OnSettingsButtonPressed()
